requirements.txt
langchain==0.3.3
langchain-core==0.3.10
langchain-community==0.3.2
openai
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
flask
python-dotenv
langchain-openai

